% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{addBaseTypeIntoRxn}
\alias{addBaseTypeIntoRxn}
\title{This function is used to give the "base" sign for two of metabolites from each reaction}
\usage{
addBaseTypeIntoRxn(rxn_split_refine_inf, metabolite_inf,
  currency_metabolites_inf)
}
\arguments{
\item{rxn_split_refine_inf}{a dataframe contains the split rxn information}

\item{metabolite_inf}{a dataframe contains the metabolite annotation information}

\item{currency_metabolites_inf}{a vector contains the currency metabolite information}
}
\description{
This function is used to give the "base" sign for two of metabolites from each reaction
}
