% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{chooseRxnFromSubsystem}
\alias{chooseRxnFromSubsystem}
\title{This function is used to choose the reaction based on subsytem definition}
\usage{
chooseRxnFromSubsystem(rxn_split_refine_inf, subsystem0)
}
\arguments{
\item{rxn_split_refine_inf}{a dataframe contains the rxn split format with the detailed annotation information}

\item{subsystem0}{a string of defined subsystem}
}
\description{
This function is used to choose the reaction based on subsytem definition
}
