% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{chooseRxnFromSubsystem_new}
\alias{chooseRxnFromSubsystem_new}
\title{Title}
\usage{
chooseRxnFromSubsystem_new(rxn_split_refine_inf, subsystem0)
}
\arguments{
\item{subsystem0}{}
}
\description{
Title
}
