% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_using_R.R
\name{creatRedge}
\alias{creatRedge}
\title{function to define the edge}
\usage{
creatRedge(rxn_core_carbon_inf, rxn_rev_list, node_inf)
}
\arguments{
\item{rxn_core_carbon_inf}{a dataframe contains the detailed annotation of rxn_core}

\item{rxn_rev_list}{a vector contains the reversible reactions}

\item{node_inf}{a dataframe contains the node definition for a R graph}
}
\description{
function to define the edge
}
