% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_using_R.R
\name{creatRmap}
\alias{creatRmap}
\title{function to creat the R graph based on the R platform}
\usage{
creatRmap(nodes, edges)
}
\arguments{
\item{nodes}{a dataframe contains the id and label information of each node}

\item{edges}{a dataframe defining the connection of the metabolite}
}
\description{
function to creat the R graph based on the R platform
}
