% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_using_R.R
\name{creatRnode}
\alias{creatRnode}
\title{function to define the node in a R graph}
\usage{
creatRnode(rxn_core_carbon_inf, currency_metabolites_inf)
}
\arguments{
\item{rxn_core_carbon_inf}{a dataframe contains the detailed annotation of rxn_core}

\item{currency_metabolites_inf}{a vector of currency metabolites}
}
\description{
function to define the node in a R graph
}
