% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{DefineAllMet}
\alias{DefineAllMet}
\title{The function is used to define the met in SBML file}
\usage{
DefineAllMet(specie, id, Metaid)
}
\arguments{
\item{specie}{specie id for the metabolite}

\item{id}{id for the metabolite}

\item{Metaid}{meta id for the metabolite}
}
\description{
The function is used to define the met in SBML file
}
