% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{DefineBaseMetabolite}
\alias{DefineBaseMetabolite}
\title{Define the base reactant and product for cellDesigner}
\usage{
DefineBaseMetabolite(rxnID0, rxn_split_refine_inf)
}
\arguments{
\item{rxnID0}{a reaction ID}

\item{rxn_split_refine_inf}{a datafram contains the relation of rxnID0 and metabolite, as well as metabolite formula and and carbon number}
}
\description{
Define the base reactant and product for cellDesigner
}
