% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{DefineCurrencyMet}
\alias{DefineCurrencyMet}
\title{This function is used to define the currency of metabolites in the model}
\usage{
DefineCurrencyMet(rxn_split_refine0, subsystem0, numberGEM,
  numberSubsystem)
}
\arguments{
\item{rxn_split_refine0}{a dataframe contains the split rxn format with detailed annotation in 10 columns}

\item{subsystem0}{a subsytem string}

\item{numberGEM}{a number used to limit the number of currency metabolite from the whole model}

\item{numberSubsystem}{a number used to limit the number of currency metabolite from each chose subsystems}
}
\description{
This function is used to define the currency of metabolites in the model
}
