% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_refine_and_omics_mapping.R
\name{defineGeneFoldColor}
\alias{defineGeneFoldColor}
\title{function to define the color based on the fold change}
\usage{
defineGeneFoldColor(omic_fold, up = 2, down = 0.5)
}
\arguments{
\item{omic_fold}{a vector contains the fold change values of omics data}

\item{up}{a value define the significantly increased fold changes}

\item{down}{a value define the significantly decreased fold changes}
}
\description{
function to define the color based on the fold change
}
