% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{DefineProtein}
\alias{DefineProtein}
\title{The function is used to define the general ID information of gene and protein in SMBL file}
\usage{
DefineProtein(p_specie, p_id, p_metaid, g_specie, g_id, g_metaid)
}
\arguments{
\item{p_specie}{specie id for protein}

\item{p_id}{id for protein}

\item{p_metaid}{meta id for protein}

\item{g_specie}{specia id for gene}

\item{g_id}{id for gene}

\item{g_metaid}{meta id for gene}
}
\description{
The function is used to define the general ID information of gene and protein in SMBL file
}
