% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{DeProteinConnect}
\alias{DeProteinConnect}
\title{The function is used to define the connection of general gene id and protein id with each reaction}
\usage{
DeProteinConnect(p_specie, p_id, g_specie, g_id)
}
\arguments{
\item{p_specie}{specie id for the protein}

\item{p_id}{id for the protein}

\item{g_specie}{specie id for the gene}

\item{g_id}{id for the gene}
}
\description{
The function is used to define the connection of general gene id and protein id with each reaction
}
