% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_refine_and_omics_mapping.R
\name{fluxMapping}
\alias{fluxMapping}
\title{Title}
\usage{
fluxMapping(input_map, flux_inf, output_map)
}
\arguments{
\item{input_map}{diretory of xml file}

\item{flux_inf}{a dataframe contains the flux value for each reaction}

\item{output_map}{diretory of xml file}
}
\description{
Title
}
