% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Define annotation information for each metabolite}
\usage{
getAnnotation(species01, name01)
}
\arguments{
\item{species01}{species id of a component}

\item{name01}{name of a component}
}
\description{
Define annotation information for each metabolite
}
