% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{getCompositionNum}
\alias{getCompositionNum}
\title{The function is used to extract the number of carbon based on the above function}
\usage{
getCompositionNum(MET, type = "C")
}
\arguments{
\item{MET}{a carbon part of a molecular, like C5}

\item{type}{the string of atom kind, like "C"}
}
\description{
The function is used to extract the number of carbon based on the above function
}
