% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{getConnectedTransport}
\alias{getConnectedTransport}
\title{Title}
\usage{
getConnectedTransport(id, rxn_transport_id0, rxn_transport0,
  met_core_carbon0)
}
\arguments{
\item{met_core_carbon0}{}
}
\description{
Title
}
