% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{getGeneAnnotation}
\alias{getGeneAnnotation}
\title{Define annotation information for each gene}
\usage{
getGeneAnnotation(species01, name01, geneID01)
}
\arguments{
\item{species01}{specie id of gene}

\item{name01}{name of a component}

\item{geneID01}{ID of the gene as the component}
}
\description{
Define annotation information for each gene
}
