% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{getMetaboliteComposition}
\alias{getMetaboliteComposition}
\title{The fuction is used to extract the carbon part for a metabolite}
\usage{
getMetaboliteComposition(MET, type = "C")
}
\arguments{
\item{MET}{a metabolite formula, like C5H6}

\item{type}{the string of atom kind, like "C"}
}
\description{
The fuction is used to extract the carbon part for a metabolite
}
