% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{getMetType}
\alias{getMetType}
\title{This function is used to estimate the metabolite type in a reaction.}
\usage{
getMetType(rxnID_inf, rxn_split_inf)
}
\arguments{
\item{rxnID_inf}{a vector contains the reaction id}

\item{rxn_split_inf}{a dataframe with three columns contains the splited rxn information}
}
\description{
This function is used to estimate the metabolite type in a reaction.
}
