% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{getNewSize}
\alias{getNewSize}
\title{Define the size of graph}
\usage{
getNewSize(x = 600, y = 400)
}
\arguments{
\item{x}{a value determining the length of graph}

\item{y}{a value determining the height of graph}
}
\description{
Define the size of graph
}
