% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{getOtherProduct}
\alias{getOtherProduct}
\title{The function is used to define the other product (except the base product) in a reaction}
\usage{
getOtherProduct(P_specie, P_id)
}
\arguments{
\item{P_specie}{specie id for the product}

\item{P_id}{id for the product}
}
\description{
The function is used to define the other product (except the base product) in a reaction
}
