% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{getOtherReactant}
\alias{getOtherReactant}
\title{The output of the followed function could be replaced of R2
The function is used to define the other reactant (except the base reactant) in a reaction}
\usage{
getOtherReactant(R_specie, R_id)
}
\arguments{
\item{R_specie}{specie id for the reactant}

\item{R_id}{id for the reactant}
}
\description{
The output of the followed function could be replaced of R2
The function is used to define the other reactant (except the base reactant) in a reaction
}
