% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{getPosition}
\alias{getPosition}
\title{Define position information for each metabolite}
\usage{
getPosition(id01, species01, x01, y01)
}
\arguments{
\item{id01}{id of a component}

\item{species01}{species id of a component}

\item{x01}{x coordinate of a component}

\item{y01}{y coordinate of a component}
}
\description{
Define position information for each metabolite
}
