% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{getProAnnotation}
\alias{getProAnnotation}
\title{Define annotation information for each protein}
\usage{
getProAnnotation(species01, name01, proteinID01)
}
\arguments{
\item{species01}{species id of protein}

\item{name01}{name of a component}

\item{proteinID01}{ID of the protein as the component}
}
\description{
Define annotation information for each protein
}
