% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{getReactionMain}
\alias{getReactionMain}
\title{The output of the followed function could be replaced of R1
This function is used to define the base reactant and production information for each reaction}
\usage{
getReactionMain(rxnID, BR_specie, BR_id, BP_specie, BP_id)
}
\arguments{
\item{BR_specie}{specie id for the base reactant}

\item{BR_id}{id for the base reactant}

\item{BP_specie}{specie id for the base product}

\item{BP_id}{id for the base product}
}
\description{
The output of the followed function could be replaced of R1
This function is used to define the base reactant and production information for each reaction
}
