% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{getRxnInformation}
\alias{getRxnInformation}
\title{Define the reaction metaid, baseReactant and baseProduct in cellDesigner}
\usage{
getRxnInformation(rxn0, rxn_p, R30 = R3, R70 = R7)
}
\arguments{
\item{rxn0}{a dataframe contains the rxn annotation}

\item{rxn_p}{a dataframe contains the general id information of gene and protein for each reaction}

\item{R30}{the template file of cellDesigner file}

\item{R70}{the template file of cellDesigner file}
}
\description{
Define the reaction metaid, baseReactant and baseProduct in cellDesigner
}
