% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_refine_and_omics_mapping.R
\name{omicsFoldMapping}
\alias{omicsFoldMapping}
\title{Function to mapping the omics fold changes onto the map}
\usage{
omicsFoldMapping(input_map, flux_inf, output_map)
}
\arguments{
\item{input_map}{diretory of xml file}

\item{flux_inf}{a dataframe contains the omic(gene or protein) fold changes value for each reaction under two different conditions}

\item{output_map}{diretory of xml file}
}
\description{
Function to mapping the omics fold changes onto the map
}
