% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{prepareMET}
\alias{prepareMET}
\title{Title}
\usage{
prepareMET(rxn_core_carbon_inf, currency_metabolites_inf, rxnID_choose_inf,
  x_vector = seq(100, 12000, by = 100), y_vector = seq(100, 15000, by =
  50))
}
\arguments{
\item{y_vector}{}
}
\description{
Title
}
