% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{prepareRXN}
\alias{prepareRXN}
\title{Title}
\usage{
prepareRXN(rxn_core_carbon_inf, met_annotation_inf,
  currency_metabolites_inf)
}
\arguments{
\item{currency_metabolites_inf}{}
}
\description{
Title
}
