% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition for cellDesigner.R
\name{produceInputForCellDesigner}
\alias{produceInputForCellDesigner}
\title{The function is used to produce the cell designer file}
\usage{
produceInputForCellDesigner(met_annotation_inf, gpr_inf,
  rxn_core_carbon_inf, x_size = 2500, y_size = 3600)
}
\arguments{
\item{met_annotation_inf}{a dataframe contains the met annotation of the chose reaction}

\item{gpr_inf}{a dataframe contains the GPR information of the chose reaction}

\item{rxn_core_carbon_inf}{a dataframe contains the detailed rxn annotation}

\item{x_size}{a number define the width of graph}

\item{y_size}{a number define the height of graph}
}
\description{
The function is used to produce the cell designer file
}
