% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{removeRxnWithSingleMet}
\alias{removeRxnWithSingleMet}
\title{When removing the currency metabolites. some reaction could cotain only metabolite
This function is used to remove the exchange reaction and reactions with only one reactant or product}
\usage{
removeRxnWithSingleMet(rxn_split = rxn_split_refine)
}
\arguments{
\item{rxn_split}{a dataframe contains the splited rxn}
}
\description{
When removing the currency metabolites. some reaction could cotain only metabolite
This function is used to remove the exchange reaction and reactions with only one reactant or product
}
