% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{rxnGeneMapping}
\alias{rxnGeneMapping}
\title{This function is used to establish the mappping between rxn and gene
Title}
\usage{
rxnGeneMapping(rxnid_gpr)
}
\arguments{
\item{rxnid_gpr}{a dataframe contains two columns, GPR and Abbreviation}
}
\description{
This function is used to establish the mappping between rxn and gene
Title
}
