% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{splitAndCombine0}
\alias{splitAndCombine0}
\title{This function is used to establish the single mapping between one gene and one rxn}
\usage{
splitAndCombine0(gene, rxn, sep0)
}
\arguments{
\item{gene}{a vector contains GPR relation for each reaction}

\item{rxn}{a vector contains the rxnid}

\item{sep0}{a sign to connect the gene}
}
\description{
This function is used to establish the single mapping between one gene and one rxn
}
