% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{splitRxnToMetabolite}
\alias{splitRxnToMetabolite}
\title{This function is used to split the reaction into the metabolite format}
\usage{
splitRxnToMetabolite(reationFrame, sep0)
}
\arguments{
\item{reationFrame}{dataframe: column 1--ID0; column 2--Equation, which is the reaction formula}

\item{sep0}{the string to connect the reactant and product in a reaction}
}
\description{
This function is used to split the reaction into the metabolite format
}
