% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{splitRxnToMetabolite.Yeast}
\alias{splitRxnToMetabolite.Yeast}
\title{This function is designed for yeast GEM to split the reaction}
\usage{
splitRxnToMetabolite.Yeast(rxn_inf, metabolite_inf)
}
\arguments{
\item{rxn_inf}{a dataframe contains the rxn annotation of yeast GEM}

\item{metabolite_inf}{a dataframe contains the metabolite annotation of yeast GEM}
}
\description{
This function is designed for yeast GEM to split the reaction
}
