% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{addBaseTypeIntoRxn}
\alias{addBaseTypeIntoRxn}
\title{Title}
\usage{
addBaseTypeIntoRxn(rxn_split_refine_inf, metabolite_inf,
  currency_metabolites_inf)
}
\arguments{
\item{currency_metabolites_inf}{}
}
\description{
Title
}
