% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_refine_and_omics_mapping.R
\name{changeFontSize}
\alias{changeFontSize}
\title{Title}
\usage{
changeFontSize(positionID, uniqueID, onemap)
}
\arguments{
\item{positionID}{the position of the component}

\item{uniqueID}{the unique ID of the metabolite}

\item{onemap}{a map in xml format}
}
\description{
Title
}
