% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model change.R
\name{getConnectedTransport_new}
\alias{getConnectedTransport_new}
\title{Title}
\usage{
getConnectedTransport_new(rxn_split_refine_inf0, met_core_carbon0)
}
\arguments{
\item{met_core_carbon0}{}
}
\description{
Title
}
